Extract the Net module (networking/SSRF/URL primitives) into Api/Net
First module of the #777 folder migration (folders WITH matching namespaces,
option B). Moves the low-level, dependency-free networking primitives into
Jellyfin.Plugin.SSO_Auth.Api.Net so the module is visible in the tree and in
every using:

- IpAddressClassifier, CanonicalBaseUrl, SsoHttp -> Api/Net/ (namespace
  Api.Net), with the import added at each call site.

SsoUrlBuilder deliberately stays in the flat Api namespace: it is not a
networking leaf — it composes SSO callback/ACS URLs and calls into
OidcCallbackPath, so it belongs with the routing/URL types that will form a
later module. The migration surfacing that coupling is the point of doing this
one module at a time.

Locks the boundary in with a new conformance fitness function,
NetModule_IsALeaf_ImportsNoOtherApiModule: no source file under Api/Net may
import another Api module, enforced at the import level so it also catches
method-body coupling (which signature reflection would miss). Each further
module lands its own allowed-direction rule, together pinning the module DAG.

Pure mechanical move: no behavior change. Verified by the full build
(--warnaserror, both TFMs), the complete test suite (1572 both TFMs), every
existing conformance rule, and the ABI-floor build.

Refs #777
