+++ SSO-Auth.Tests/ArchitectureConformanceTests.cs
@@ -148,6 +148,28 @@ public class ArchitectureConformanceTests
     }
 
     [Fact]
+    public void NetModule_IsALeaf_ImportsNoOtherApiModule()
+    {
+        // First module-boundary fitness function of the #777 folder migration. The Net module holds the
+        // low-level networking / URL / SSRF primitives (IpAddressClassifier, CanonicalBaseUrl, SsoHttp); it is
+        // a LEAF — dependencies point INTO it, never out. Enforced at the IMPORT level, which also catches
+        // method-body coupling (reflection over signatures would miss a body-only call): using any type from
+        // another Api module requires importing its namespace, so no source file under SSO-Auth/Api/Net may
+        // import a Jellyfin.Plugin.SSO_Auth.Api or Jellyfin.Plugin.SSO_Auth.Api.* namespace. As further modules
+        // land (#777), each adds its own allowed-direction rule; together they lock in the module DAG.
+        var netDir = Path.Combine(RepoRoot(), "SSO-Auth", "Api", "Net");
+        var offenders = Directory.EnumerateFiles(netDir, "*.cs")
+            .SelectMany(file => File.ReadLines(file)
+                .Where(line => Regex.IsMatch(line, @"^\s*using\s+Jellyfin\.Plugin\.SSO_Auth\.Api(\.|;)"))
+                .Select(line => Path.GetFileName(file) + ": " + line.Trim()))
+            .ToList();
+
+        Assert.True(
+            offenders.Count == 0,
+            "The Net module is a leaf: no file under Api/Net may import another Api module. Offending imports: " + string.Join(" | ", offenders));
+    }
+
+    [Fact]
     public void MutableKeyedState_LivesOnlyInsideStoreLikeTypes()
     {
         // Locked in by the OidcStateStore consolidation (#318): a raw dictionary holding runtime state
