+++ SSO-Auth/Api/ProviderNameValidator.cs
@@ -2,6 +2,7 @@
 
 using System;
 using System.Buffers;
+using Jellyfin.Plugin.SSO_Auth.Api.Net;
 
 namespace Jellyfin.Plugin.SSO_Auth.Api;
